Initial development to reduce fixed encounter RNG, such that each runner in a seed encounters the same fixed encounters. Added documentation to better cross-reference with FF6Tools and set up every Fixed Encounter to be identical regardless of RNG.
