+++ data/enemy_formation.py
@@ -1,3 +1,4 @@
+# Within https://everything8215.github.io/ff6tools/ff6tools.html, these are what's found under Battle -> Battles
 class EnemyFormation():
     ENEMY_CAPACITY = 6 # up to 6 enemies in a formation
